requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pdfplumber==0.11.6
python-docx==1.1.2
scikit-learn==1.4.2
spacy==3.8.5
streamlit==1.32.0
